refactor(main.js, data.js, utils.js): create modules
