+++ js/utils.js
@@ -1,0 +1,5 @@
+const getRandomNumber = (min, max) => (Math.floor(Math.random() * (max - min) + min));
+
+const checkStringLength = (str, maxLength) => (str.length <= maxLength);
+
+export {getRandomNumber, checkStringLength};
